+++ package.json
@@ -27,7 +27,7 @@
     "koa-bodyparser": "^4.3.0",
     "koa-router": "^10.1.1",
     "lodash": "^4.17.21",
-    "mongoose": "^6.0.13",
+    "mongoose": "^6.0.14",
     "mongoose-findorcreate": "^3.0.0"
   },
   "devDependencies": {
@@ -39,7 +39,7 @@
     "@types/koa-router": "^7.4.4",
     "@types/koa__cors": "^3.0.3",
     "@types/lodash": "^4.14.177",
-    "@types/node": "^16.11.10",
+    "@types/node": "^16.11.11",
     "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^5.5.0",
     "@typescript-eslint/parser": "^5.5.0",
@@ -51,6 +51,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
+    "http-graceful-shutdown": "^3.1.5",
     "jest": "^27.4.0",
     "module-alias": "^2.2.2",
     "mongodb-memory-server": "^8.0.3",
