+++ src/__tests__/login.ts
@@ -1,40 +1,47 @@
 import * as request from 'supertest'
+import * as shutdown from 'http-graceful-shutdown'
 import { MongoMemoryServer } from 'mongodb-memory-server'
+import { Mongoose } from 'mongoose'
 import { Server } from 'http'
-import { dropMongo, startKoa, stopServer } from '@/__tests__/testUtils'
-import { runMongo, stopMongo } from '@/helpers/mongo'
 import MockAdapter from 'axios-mock-adapter'
-import app from '@/app'
 import axios from 'axios'
-import testingGoogleMock from '@/helpers/testingGoogleMock'
+import runApp from '@/helpers/runApp'
+import runMongo from '@/helpers/mongo'
 
 describe('Login endpoint', () => {
   const axiosMock = new MockAdapter(axios)
+
   let server: Server
+  let mongoServer: MongoMemoryServer
+  let mongoose: Mongoose
 
   beforeAll(async () => {
-    const mongoServer = await MongoMemoryServer.create()
-    await runMongo(await mongoServer.getUri())
-    server = await startKoa(app)
+    mongoServer = await MongoMemoryServer.create()
+    mongoose = await runMongo(await mongoServer.getUri())
+    server = await runApp()
   })
 
   beforeEach(async () => {
-    await dropMongo()
+    await mongoose.connection.db.dropDatabase()
   })
 
   afterAll(async () => {
-    await stopMongo()
-    await stopServer(server)
+    await shutdown(server)
+    await mongoServer.stop()
   })
 
   it('should return user for valid /google request', async () => {
+    const testingGoogleMock = {
+      name: 'John Doe',
+      email: 'john@doe.com',
+    }
     axiosMock
       .onGet('https://www.googleapis.com/oauth2/v3/tokeninfo?id_token=test')
       .reply(200, testingGoogleMock)
     const response = await request(server)
       .post('/login/google')
       .send({ accessToken: 'test' })
-
+    console.log(response.error)
     expect(response.body.name).toBe(testingGoogleMock.name)
     expect(response.body.email).toBe(testingGoogleMock.email)
   })
