--- src/app.ts
@@ -1,28 +1,0 @@
-import 'reflect-metadata'
-import * as Koa from 'koa'
-import * as Router from 'koa-router'
-import * as bodyParser from 'koa-bodyparser'
-import * as cors from '@koa/cors'
-import { bootstrapControllers } from 'amala'
-
-const app = new Koa()
-void (async () => {
-  try {
-    const router = new Router()
-    await bootstrapControllers({
-      app,
-      router,
-      basePath: '/',
-      controllers: [__dirname + '/controllers/*'],
-      disableVersioning: true,
-    })
-    app.use(cors({ origin: '*' }))
-    app.use(bodyParser())
-    app.use(router.routes())
-    app.use(router.allowedMethods())
-  } catch (err) {
-    console.log('Koa app starting error: ', err)
-  }
-})()
-
-export default app
