+++ src/helpers/mongo.ts
@@ -1,13 +1,6 @@
 import * as mongoose from 'mongoose'
 import env from '@/helpers/env'
 
-export function runMongo(mongoUrl = env.MONGO) {
-  if (!mongoUrl) {
-    throw new Error('MONGO is not defined')
-  }
+export default function runMongo(mongoUrl = env.MONGO) {
   return mongoose.connect(mongoUrl)
-}
-
-export function stopMongo() {
-  return mongoose.disconnect()
 }
