--- src/helpers/testingGoogleMock.ts
@@ -1,6 +1,0 @@
-const testingGoogleMock = {
-  name: 'John Doe',
-  email: 'john@doe.com',
-}
-
-export default testingGoogleMock
